workflow-analyzer-saas/runner/requirements.txt
google-generativeai>=0.4.0
opencv-python>=4.10.0
rich>=13.7
python-dotenv>=1.0
tenacity>=8.2
ffmpeg-python>=0.2
supabase>=2.0.0
aiohttp>=3.9.0
reportlab>=4.0.0
Pillow>=10.0.0       
